refactor: move DockerSandboxValidator to deployments module
